+++ src/app/components/query-builder/block-settings.js
@@ -41,8 +41,8 @@ export default class QueryBuilderBlockSettingsComponent extends Component {
   }
 
   @action
-  popoverVisibilityChange(isShown) {
-    if (!isShown && this.isShown) {
+  popoverWillHide() {
+    if (this.isShown) {
       this.onSettingsClose();
     }
   }
@@ -54,13 +54,6 @@ export default class QueryBuilderBlockSettingsComponent extends Component {
   @action
   blockReplace(closeSelectorCallback, newBlocks) {
     closeSelectorCallback();
-
-    // Immediate notify about close and not wait for animation.
-    // Fixes issue with reopening block settings after removing block via "None" operator.
-    // In that situation parent was not notified about closed selector, because
-    // event origin block was destroyed before hiding selector animation end.
-    this.onSettingsClose();
-
     this.onBlockReplace(newBlocks);
   }
 }